feat: add eslint cypress plugin
Add eslint to package.json dev-dependencies.

Append plugin:cypress/recommended to end of extends array within the
eslintConfig object from package.json.

Extract the eslintConfig from package.json file and append it to a new
.eslintrc.json file for better locating.

Important to not that this commit is dependent on the
eslint-plugin-cypress dev-dependency added from the first commit.